feat: move ui-colors to react
