--- packages/ui/colors/.prettierignore
@@ -1,7 +1,0 @@
-package-lock.json
-yarn.lock
-node_modules
-*.yml
-.eslintignore
-.prettierignore
-.gitignore
