--- packages/ui/colors/.prettierrc.json
@@ -1,13 +1,0 @@
-{
-  "$schema": "http://json.schemastore.org/prettierrc",
-  "plugins": [
-    "prettier-plugin-multiline-arrays"
-  ],
-  "tabWidth": 2,
-  "semi": false,
-  "singleQuote": true,
-  "printWidth": 80,
-  "trailingComma": "es5",
-  "multilineArraysWrapThreshold": 0,
-  "singleAttributePerLine": true
-}
