--- packages/ui/colors/tsconfig.json
@@ -1,33 +1,0 @@
-{
-  "compilerOptions": {
-    /* Basic Options */
-    "target": "es2015",
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "lib": [
-      "ES2015",
-      "DOM"
-    ],
-    "allowJs": false,
-    "declaration": true,
-    "resolveJsonModule": true,
-    "rootDir": "./src",
-    "outDir": "./dist",
-    "removeComments": false,
-    /* Strict Type-Checking Options */
-    "strict": true,
-    "noImplicitAny": true,
-    "noUnusedLocals": false,
-    "strictNullChecks": true,
-    /* Module Resolution Options */
-    "esModuleInterop": true,
-    /* Advanced Options */
-    "forceConsistentCasingInFileNames": true
-  },
-  "include": [
-    "src/**/*"
-  ],
-  "exclude": [
-    "tests/*"
-  ]
-}
