+++ packages/ui/react/src/colors/index.ts
@@ -1,0 +1,2 @@
+export * from './getColor'
+export * from './types'
